COMP: Fix `elx::Conversion` "warning: '==': signed/unsigned mismatch"
Fixed:
> warning C4389: '==': signed/unsigned mismatch

Encountered in `elx::Conversion::StringToValue` when doing MSVC 2022 compilation with "/W4".

+++ Core/Install/elxConversion.h
@@ -190,7 +190,8 @@ public:
       const auto decimalPointPos = str.find_first_of('.');
       const bool hasDecimalPointAndTrailingZeros =
         (decimalPointPos != std::string::npos) &&
-        (std::count(str.cbegin() + decimalPointPos + 1, str.cend(), '0') == (str.size() - decimalPointPos - 1));
+        (static_cast<std::uintmax_t>(std::count(str.cbegin() + decimalPointPos + 1, str.cend(), '0')) ==
+         (str.size() - decimalPointPos - 1));
       return std::istringstream(
         hasDecimalPointAndTrailingZeros ? std::string(str.cbegin(), str.cbegin() + decimalPointPos) : str);
     }();
